Added log-file, log_file configuration option.
The log-file specifies the file to log to. The actual log
goes to log-file.current and on SIGHUP, the .current file gets
copied to the log-file and then the .current is truncated. This is
to allow logrotate like commands to be able to rotate consul logs.

For example, log-file=/var/log/consul.log will log to /var/log/consul.log.current
and on SIGHUP, will copy the file to /var/log/consul.log and then
truncate /var/log/consul.log.current without affecting the *File so that
consul can keep logging.